\docType{package}
\name{hdxdictionary-package}
\alias{hdxdictionary-package}
\title{Convert country codes or country names}
\description{
Convert country codes or country names
}
\details{
The \code{hdxdictionary}

It was forked from the \code{countrycode} package.
}
\author{
Vincent Arel-Bundock \email{varel@umich.edu}
}
\references{
\url{http://www.umich.edu/~varel}
\url{https://github.com/vincentarelbundock/countrycode}
}
\keyword{package}

